+++ tests/writing/test_serialize_party.py
@@ -9,7 +9,7 @@ from textwrap import dedent
 
 import pytest
 
-from lanpartydb.models import Links, Location, Party, Resource
+from lanpartydb.models import Location, Party, PartyLinks, Resource
 from lanpartydb.writing import serialize_party
 
 
@@ -50,7 +50,7 @@ from lanpartydb.writing import serialize_party
                     latitude=Decimal('34.06101057935884'),
                     longitude=Decimal('-118.23974355902666'),
                 ),
-                links=Links(
+                links=PartyLinks(
                     website=Resource(
                         url='https://www.superlan.example/',
                         offline=False,