+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## 0.11.0 (unreleased)
 
+- Renamed model `Links` to `PartyLinks`.
+
 - Updated pytest to v9.0.2 (from v8.3.5).
 
 - Updated ruff to v0.14.8 (from v0.11.3).